requirements.txt
Cython
matplotlib
numpy
omegaconf>=2.0.1
onnx
onnx-simplifier
opencv-python
pyaml
pycocotools
pytorch-lightning>=1.4.0
tabulate
tensorboard
termcolor
# torch>=1.7
torchmetrics
tqdm
torch==1.7.1+cu110 -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.8.2+cu110 -f https://download.pytorch.org/whl/torch_stable.html
# torchaudio==0.9.1 -f https://download.pytorch.org/whl/torch_stable.html